requirements.txt
# Core Django
Django==5.0
django-environ==0.11.2

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS (Frontend için)
django-cors-headers==4.3.1

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Caching
redis>=4.5.2,<5.0.0
django-redis==5.4.0

# Async Tasks
celery[redis]==5.3.4

# Image Processing (profile pictures için)
Pillow==10.4.0

# API requests (TMDB API için)
requests==2.31.0

# Environment variables
python-decouple==3.8

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# AI/ML - Recommendation System
scikit-learn>=1.3.0
scipy>=1.11.0
torch>=2.0.0
numpy>=1.24.0

# Code Quality
flake8==6.1.0
black==23.12.0

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Azure Integration
azure-storage-blob==12.19.0

# Additional
django-filter==23.5